타이머 임시 저장
