--- Assets/PlayFabEditorExtensions/Editor/Resources/MostRecentPackage.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 56dc732589f2c48448ce602fd5931c8e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
