+++ Assets/Scripts/Player/Player.cs
@@ -10,8 +10,6 @@ public class Player : MonoBehaviour
     [SerializeField] UIManager uiManager;
     private PhotonView pv;
 
-    int time;
-
     // Start is called before the first frame update
     void Awake()
     {
@@ -48,39 +46,5 @@ public class Player : MonoBehaviour
         {
             Debug.Log("points가 왜 null임??");
         }
-    }
-
-    void StartTimer()
-    {
-        if (PhotonNetwork.IsMasterClient)
-        {
-            time = 60;
-
-            StartCoroutine(TimerCoroution());
-        }
-    }
-
-    IEnumerator TimerCoroution()
-    {
-        if (time > 0)
-        {
-            time -= 1;
-        }
-        else
-        {
-            Debug.Log("타이머 종료");
-            yield break;
-        }
-
-        pv.RPC("ShowTimer", RpcTarget.All, time); //1초 마다 방 모두에게 전달
-
-        yield return new WaitForSeconds(1);
-        StartCoroutine(TimerCoroution());
-    }
-
-    [PunRPC]
-    void ShowTimer(int time)
-    {
-        Debug.Log(time);
     }
 }
