+++ Assets/Scripts/System/Timer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d40df4a2e329274d989157cbbc5fe05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
